@@querydsl환경설정 최종@@ fix) 게시글 조회가 안됐던 원인fixed add)게시글이 없을 시 msg 추가

+++ Backend/src/main/java/com/ssafy/common/domain/problem/Problem_SitePK.java
@@ -8,5 +8,5 @@ import lombok.Data;
 public class Problem_SitePK implements Serializable{
   // private static final long serialVersionUID = -3033073092220146757L;
   private String problemSiteName;
-  private int problemNo;
+  private long problemNo;
 }
